@@ -1,18 +1,21 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
     <LangVersion>latest</LangVersion>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="FluentValidation.AspNetCore" Version="10.3.3" />
+    <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="11.3.0" />
+    <PackageReference Include="NSwag.AspNetCore" Version="13.17.0" />
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.*" />
     <PackageReference Include="Swashbuckle.AspNetCore.Annotations" Version="6.*" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\SI.Endpoints.NSwag\SI.Endpoints.NSwag.csproj" />
     <ProjectReference Include="..\SI.Endpoints.Swagger\SI.Endpoints.Swagger.csproj" />
   </ItemGroup>
 
